empty back accounts design added

+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/fragments/BankAccountFragment.kt
@@ -57,8 +57,14 @@ class BankAccountFragment : Fragment() {
 
 
 
-        binding.fab.setOnClickListener {
+        binding.addFab.setOnClickListener {
+            if(binding.addFab.visibility == View.VISIBLE)
             view?.findNavController()?.navigate(actionBankAccountFragmentToAddCreditCardFragment())
+        }
+
+        binding.btnAddCard.setOnClickListener {
+            if(binding.btnAddCard.visibility == View.VISIBLE)
+                view?.findNavController()?.navigate(actionBankAccountFragmentToAddCreditCardFragment())
         }
 
         val credits = arrayListOf<CardModel>()
@@ -74,20 +80,34 @@ class BankAccountFragment : Fragment() {
 
         viewModel.creditList.observe(viewLifecycleOwner, Observer { list ->
             list?.let {
-                val creditCartsAdapter = CreditCartsAdapter(ArrayList(list!!),this)
-                recView.adapter = creditCartsAdapter
-                recView.setHasFixedSize(true)
-                // creditCartsAdapter.submitList(it)
-                val itemTouchHelper = ItemTouchHelper(SwipeToDeleteCreditCart(creditCartsAdapter!!))
-                itemTouchHelper.attachToRecyclerView(recView)
-                creditCartsAdapter.pos.observe(viewLifecycleOwner, Observer {
-                    if (it != -1) {
-                        val id = creditCartsAdapter.deleteItem(it).id
-                        viewModel.deleteCustomerCard(id)
-                       // viewModel.getCustomerCards()
-                        creditCartsAdapter.resetPos()
-                    }
-                })
+                if(it.size!=0) {
+                    binding.btnAddCard.visibility = View.GONE
+                    binding.creditImage.visibility = View.GONE
+                    binding.creditText.visibility = View.GONE
+                    binding.creditList.visibility = View.VISIBLE
+                    binding.addFab.visibility = View.VISIBLE
+                    val creditCartsAdapter = CreditCartsAdapter(ArrayList(list!!), this)
+                    recView.adapter = creditCartsAdapter
+                    recView.setHasFixedSize(true)
+                    // creditCartsAdapter.submitList(it)
+                    val itemTouchHelper =
+                        ItemTouchHelper(SwipeToDeleteCreditCart(creditCartsAdapter!!))
+                    itemTouchHelper.attachToRecyclerView(recView)
+                    creditCartsAdapter.pos.observe(viewLifecycleOwner, Observer {
+                        if (it != -1) {
+                            val id = creditCartsAdapter.deleteItem(it).id
+                            viewModel.deleteCustomerCard(id)
+                            // viewModel.getCustomerCards()
+                            creditCartsAdapter.resetPos()
+                        }
+                    })
+                } else {
+                    binding.btnAddCard.visibility = View.VISIBLE
+                    binding.creditImage.visibility = View.VISIBLE
+                    binding.creditText.visibility = View.VISIBLE
+                    binding.creditList.visibility = View.GONE
+                    binding.addFab.visibility = View.GONE
+                }
             }
         })
